% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_catalogs.R
\name{i14y_get_catalog}
\alias{i14y_get_catalog}
\title{Get a catalog by ID}
\usage{
i14y_get_catalog(id = NULL)
}
\arguments{
\item{id}{string. The catalog identifier.}
}
\value{
A list of catalog metadata. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint \verb{/catalogs/\{catalogId\}}.
}
