% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_catalogs.R
\name{i14y_get_catalog_dcat}
\alias{i14y_get_catalog_dcat}
\title{Export the DCAT representation of a catalog}
\usage{
i14y_get_catalog_dcat(id = NULL, format = "RDF")
}
\arguments{
\item{id}{string. The catalog identifier.}

\item{format}{string. One of "RDF" (default), "TTL".}
}
\value{
A character string with the raw RDF/TTL document. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint
\verb{/catalogs/\{catalogId\}/dcat/exports/\{dataFormat\}}.
}
