% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_catalogs.R
\name{i14y_get_catalog_record}
\alias{i14y_get_catalog_record}
\title{Get a single record of a catalog}
\usage{
i14y_get_catalog_record(id = NULL, recordId = NULL)
}
\arguments{
\item{id}{string. The catalog identifier.}

\item{recordId}{string. The record identifier.}
}
\value{
A list with the record metadata. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint
\verb{/catalogs/\{catalogId\}/records/\{recordId\}}.
}
