% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_catalogs.R
\name{i14y_get_catalog_records}
\alias{i14y_get_catalog_records}
\title{List records of a catalog}
\usage{
i14y_get_catalog_records(id = NULL)
}
\arguments{
\item{id}{string. The catalog identifier.}
}
\value{
A tibble of catalog records. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint \verb{/catalogs/\{catalogId\}/records}.
}
