% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_concept_all_versions.R
\name{i14y_get_concept_all_versions}
\alias{i14y_get_concept_all_versions}
\title{Get all versions of an ID}
\usage{
i14y_get_concept_all_versions(id = NULL, language = "de")
}
\arguments{
\item{id}{string. The Id of the response data.}

\item{language}{string. The language of the response data.}
}
\value{
a list
}
\description{
Get all versions of an ID
}
\examples{
i14y_get_concept_all_versions(
 id = "08d94604-e058-62a2-aa25-53f84b974201", # DV_NOGA_DIVISION
 language = "de"
)
}
