% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_list.R
\name{i14y_get_concept_export}
\alias{i14y_get_concept_export}
\title{Get the JSON export of a concept}
\usage{
i14y_get_concept_export(id = NULL)
}
\arguments{
\item{id}{string. The UUID of the concept.}
}
\value{
A list with the JSON export of the concept. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint
\verb{/concepts/\{conceptId\}/exports/json}.
}
