% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_data_element.R
\name{i14y_get_data_element}
\alias{i14y_get_data_element}
\title{Get a data element}
\usage{
i14y_get_data_element(id = NULL, language = "de")
}
\arguments{
\item{id}{string. The id of the sms used variable.}

\item{language}{string. The language of the response data.}
}
\value{
a tibble
i14y_get_data_element(
  id = "af3b5919-0012-42ac-bbdd-24b591279899",
  language = "en"
)
}
\description{
Get a data element
}
