% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_data_service.R
\name{i14y_get_data_service}
\alias{i14y_get_data_service}
\title{Get the dcat data service.}
\usage{
i14y_get_data_service(id = NULL, language = "de")
}
\arguments{
\item{id}{string. The id of the dcat data service.}

\item{language}{string. The language of the response data.}
}
\value{
a list
}
\description{
Get the dcat data service.
}
\examples{
i14y_get_data_service(
 id = "0aaed69d-15ff-415a-8a38-4f9f0c3b429e",
 language = "en"
)
}
