% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_data_service_registration.R
\name{i14y_get_data_service_registration}
\alias{i14y_get_data_service_registration}
\title{Get the current registration status of the specified dataservice.}
\usage{
i14y_get_data_service_registration(id = NULL)
}
\arguments{
\item{id}{string. The id of the dcat data service.}
}
\value{
a list
}
\description{
Get the current registration status of the specified dataservice.
}
\examples{
i14y_get_data_service_registration(
 id = "0aaed69d-15ff-415a-8a38-4f9f0c3b429e"
)
}
