% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_data_structure.R
\name{i14y_get_data_structure}
\alias{i14y_get_data_structure}
\title{Get the dataStructures}
\usage{
i14y_get_data_structure(identifier = NULL, language = "de")
}
\arguments{
\item{identifier}{string. The identifier of the dcat dataset.}

\item{language}{string. The language of the response data.}
}
\value{
a list
}
\description{
Get the dataStructures
}
\examples{
df <- i14y_get_data_structure(
  identifier = "SpiGes_Erhebung_Administratives"
)
}
