% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_dataservices.R
\name{i14y_get_dataservice}
\alias{i14y_get_dataservice}
\title{Get a data service by ID}
\usage{
i14y_get_dataservice(id = NULL)
}
\arguments{
\item{id}{string. The UUID of the data service.}
}
\value{
A list with the data service metadata. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint \verb{/dataservices/\{dataServiceId\}}.
}
