% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_dataset.R
\name{i14y_get_dataset}
\alias{i14y_get_dataset}
\title{Get dataset metadata by ID}
\usage{
i14y_get_dataset(id = NULL)
}
\arguments{
\item{id}{string. The UUID of the dataset.}
}
\value{
A list with the dataset metadata. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint \verb{/datasets/\{datasetId\}} and returns
the dataset metadata: publication level, registration status, distributions,
contact points, themes, etc.
}
\examples{
\donttest{
i14y_get_dataset(
  id = "b902add5-9538-47ed-b663-f9fbfac92381" # SpiGes_Administratives
)
}
}
