% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_dataset_distributions.R
\name{i14y_get_dataset_distributions}
\alias{i14y_get_dataset_distributions}
\title{Get dataset distributions by identifier}
\usage{
i14y_get_dataset_distributions(identifier = NULL, language = "de")
}
\arguments{
\item{identifier}{string. The identifier of the dcat dataset.}

\item{language}{string. The language of the response data.}
}
\value{
a list
}
\description{
Get dataset distributions by identifier
}
\examples{
i14y_get_dataset_distributions(identifier = "HCL_NOGA",  language = "en")
}
