% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_mappingtables.R
\name{i14y_get_mappingtable}
\alias{i14y_get_mappingtable}
\title{Get a mapping table by ID}
\usage{
i14y_get_mappingtable(id = NULL)
}
\arguments{
\item{id}{string. The UUID of the mapping table.}
}
\value{
A list with the mapping table metadata. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint \verb{/mappingtables/\{mappingTableId\}}.
}
