% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_nomenclature_level.R
\name{i14y_get_nomenclature_level}
\alias{i14y_get_nomenclature_level}
\title{Export a level of a nomenclature}
\usage{
i14y_get_nomenclature_level(
  identifier = NULL,
  language = "de",
  level = 1,
  format = "csv",
  annotations = FALSE,
  filters = NULL
)
}
\arguments{
\item{identifier}{string. The identifier of dcat dataset.}

\item{language}{string. The language of the response data.}

\item{level}{integer. The level to export. Default 1.}

\item{format}{string. The format of the export (CSV or XLSX).}

\item{annotations}{boolean. Include annotations. Default FALSE.}

\item{filters}{object. The filters}
}
\value{
a tibble
}
\description{
Export a level of a nomenclature
}
