% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_nomenclature_level_multiple.R
\name{i14y_get_nomenclature_level_multiple}
\alias{i14y_get_nomenclature_level_multiple}
\title{Export multiple levels of a nomenclature}
\usage{
i14y_get_nomenclature_level_multiple(
  identifier = NULL,
  language = "de",
  levelFrom = NULL,
  levelTo = NULL,
  format = "csv",
  annotations = NULL,
  filters = NULL
)
}
\arguments{
\item{identifier}{string. The identifier of dcat dataset.}

\item{language}{string. The language of the response data. Default "de".}

\item{levelFrom}{integer. The first level to include.}

\item{levelTo}{integer. The last level to include.}

\item{format}{string. The format of the export ("csv" or "xlsx"). Default "csv".}

\item{annotations}{boolean. Include annotations}

\item{filters}{object. The filters}
}
\value{
a tibble
}
\description{
Export multiple levels of a nomenclature
}
\examples{
i14y_get_nomenclature_level_multiple(
  identifier = "HCL_CH_ISCO_19_PROF",
  format = "csv", # read internally
  levelFrom = 1,
  levelTo = 2,
  language = "fr"
 )
}
