% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_publicservices.R
\name{i14y_get_publicservice}
\alias{i14y_get_publicservice}
\title{Get a public service by ID}
\usage{
i14y_get_publicservice(id = NULL)
}
\arguments{
\item{id}{string. The UUID of the public service.}
}
\value{
A list with the public service metadata. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint
\verb{/publicservices/\{publicServiceId\}}.
}
