% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_catalogs.R
\name{i14y_list_catalogs}
\alias{i14y_list_catalogs}
\title{List catalogs}
\usage{
i14y_list_catalogs(page = NULL, pageSize = NULL)
}
\arguments{
\item{page}{integer. The result page to return.}

\item{pageSize}{integer. The size of each result page.}
}
\value{
A tibble of catalog summaries. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint \verb{/catalogs}.
}
