% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_list.R
\name{i14y_list_concepts}
\alias{i14y_list_concepts}
\title{List concepts}
\usage{
i14y_list_concepts(
  conceptIdentifier = NULL,
  publisherIdentifier = NULL,
  version = NULL,
  publicationLevel = NULL,
  registrationStatus = NULL,
  page = NULL,
  pageSize = NULL
)
}
\arguments{
\item{conceptIdentifier}{string. Filter by concept identifier.}

\item{publisherIdentifier}{string. Filter by publisher identifier.}

\item{version}{string. Filter by version.}

\item{publicationLevel}{string. One of "Internal", "Public".}

\item{registrationStatus}{string. A registration status value.}

\item{page}{integer. The result page to return.}

\item{pageSize}{integer. The size of each result page.}
}
\value{
A tibble of concept summaries. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint \verb{/concepts}.
}
