% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_dataservices.R
\name{i14y_list_dataservices}
\alias{i14y_list_dataservices}
\title{List data services}
\usage{
i14y_list_dataservices(
  accessRights = NULL,
  dataServiceIdentifier = NULL,
  publisherIdentifier = NULL,
  publicationLevel = NULL,
  registrationStatus = NULL,
  page = NULL,
  pageSize = NULL
)
}
\arguments{
\item{accessRights}{string. Filter by access rights.}

\item{dataServiceIdentifier}{string. Filter by data service identifier.}

\item{publisherIdentifier}{string. Filter by publisher identifier.}

\item{publicationLevel}{string. One of "Internal", "Public".}

\item{registrationStatus}{string. A registration status value.}

\item{page}{integer. The result page to return.}

\item{pageSize}{integer. The size of each result page.}
}
\value{
A tibble of data service summaries. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint \verb{/dataservices}.
}
