% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_list.R
\name{i14y_list_datasets}
\alias{i14y_list_datasets}
\title{List datasets}
\usage{
i14y_list_datasets(
  accessRights = NULL,
  datasetIdentifier = NULL,
  publisherIdentifier = NULL,
  publicationLevel = NULL,
  registrationStatus = NULL,
  page = NULL,
  pageSize = NULL
)
}
\arguments{
\item{accessRights}{string. Filter by access rights.}

\item{datasetIdentifier}{string. Filter by dataset identifier.}

\item{publisherIdentifier}{string. Filter by publisher identifier.}

\item{publicationLevel}{string. One of "Internal", "Public".}

\item{registrationStatus}{string. A registration status value.}

\item{page}{integer. The result page to return.}

\item{pageSize}{integer. The size of each result page.}
}
\value{
A tibble of dataset summaries. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint \verb{/datasets}.
}
