% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_mappingtables.R
\name{i14y_list_mappingtables}
\alias{i14y_list_mappingtables}
\title{List mapping tables}
\usage{
i14y_list_mappingtables(
  mappingTableIdentifier = NULL,
  publisherIdentifier = NULL,
  version = NULL,
  codeSystemUri = NULL,
  publicationLevel = NULL,
  registrationStatus = NULL,
  page = NULL,
  pageSize = NULL
)
}
\arguments{
\item{mappingTableIdentifier}{string. Filter by mapping table identifier.}

\item{publisherIdentifier}{string. Filter by publisher identifier.}

\item{version}{string. Filter by version.}

\item{codeSystemUri}{string. Filter by code system URI.}

\item{publicationLevel}{string. One of "Internal", "Public".}

\item{registrationStatus}{string. A registration status value.}

\item{page}{integer. The result page to return.}

\item{pageSize}{integer. The size of each result page.}
}
\value{
A tibble of mapping table summaries. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint \verb{/mappingtables}.
}
