% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_publicservices.R
\name{i14y_list_publicservices}
\alias{i14y_list_publicservices}
\title{List public services}
\usage{
i14y_list_publicservices(
  publicServiceIdentifier = NULL,
  publisherIdentifier = NULL,
  publicationLevel = NULL,
  registrationStatus = NULL,
  page = NULL,
  pageSize = NULL
)
}
\arguments{
\item{publicServiceIdentifier}{string. Filter by public service identifier.}

\item{publisherIdentifier}{string. Filter by publisher identifier.}

\item{publicationLevel}{string. One of "Internal", "Public".}

\item{registrationStatus}{string. A registration status value.}

\item{page}{integer. The result page to return.}

\item{pageSize}{integer. The size of each result page.}
}
\value{
A tibble of public service summaries. `NULL` when offline.
}
\description{
Calls the I14Y public API endpoint \verb{/publicservices}.
}
