% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_search_nomenclature.R
\name{i14y_search_nomenclature}
\alias{i14y_search_nomenclature}
\title{Search within a nomenclature}
\usage{
i14y_search_nomenclature(
  identifier = NULL,
  query = NULL,
  language = "de",
  page = NULL,
  pageSize = NULL,
  filters = NULL
)
}
\arguments{
\item{identifier}{string. The identifier of dcat dataset.}

\item{query}{string. The search query.}

\item{language}{string. The language of the response data.}

\item{page}{integer. The number of the result page to return}

\item{pageSize}{integer. The size of each result page}

\item{filters}{object. The filters}
}
\value{
a list
}
\description{
Search within a nomenclature
}
