% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_codelist.R
\name{i14y_get_codelist}
\alias{i14y_get_codelist}
\title{Export a codelist}
\usage{
i14y_get_codelist(id = NULL, format = "csv")
}
\arguments{
\item{id}{string. The Id of the response data.}

\item{format}{string. The format of the export ("csv" or "json").}
}
\value{
a tibble
}
\description{
The function uses the I14Y Console API:
<https://apiconsole.i14y.admin.ch/public/v1/index.html>.
}
\examples{
i14y_get_codelist(
  id = "08d94604-e058-62a2-aa25-53f84b974201" # for DV_NOGA_DIVISION
)
}
