% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_content_information.R
\name{i14y_get_content_information}
\alias{i14y_get_content_information}
\title{Get the information of a nomenclature by identifier}
\usage{
i14y_get_content_information(identifier = NULL)
}
\arguments{
\item{identifier}{string. The identifier of the nomenclature.}
}
\value{
a list
}
\description{
Get the information of a nomenclature by identifier
}
\examples{
i14y_get_content_information(
  identifier = "HCL_CH_ISCO_19_PROF"
)
}
