% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_get_dataset_metadata.R
\name{i14y_get_dataset_metadata}
\alias{i14y_get_dataset_metadata}
\title{Get dataset full metadata by ID}
\usage{
i14y_get_dataset_metadata(id = NULL)
}
\arguments{
\item{id}{string. The id of the dcat dataset.}
}
\value{
a list
}
\description{
Get dataset full metadata by ID
}
\examples{
i14y_get_dataset_description(identifier = "HCL_NOGA")
}
