% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i14y_search_concept.R
\name{i14y_search_concept}
\alias{i14y_search_concept}
\title{Search for a Concept Summary}
\usage{
i14y_search_concept(
  query = NULL,
  language = "de",
  page = 1,
  pageSize = 1000,
  publishers = NULL,
  themes = NULL,
  conceptValueTypes = NULL,
  registrationStatuses = NULL
)
}
\arguments{
\item{query}{string. Search query.}

\item{language}{string. The language of the response data.}

\item{page}{integer. The number of the result page to return.}

\item{pageSize}{integer. The size of each result page.}

\item{publishers}{vector of strings. Filter with Publishers identifiers.}

\item{themes}{vector of strings. Filter with theme codes.}

\item{conceptValueTypes}{character vector. One or more concept value types to filter by.}

\item{registrationStatuses}{character vector. One or more registration statuses to filter by.}
}
\value{
a tibble
}
\description{
Search for a Concept Summary
}
\examples{
i14y_search_concept(query = "noga", language = "en")
}
